cleaned up modal files
