+++ client/src/components/LoginModal/LoginModal.css
@@ -78,3 +78,4 @@
   -webkit-text-stroke-width: 3px;
   -webkit-text-stroke-color: #363636;
 }
+
